pkg/nlp/requirements.txt
toml
requests
tornado
pynsq
transformers
gensim
Sastrawi
seaborn
scikit-learn
nltk
pdfminer.six
pymupdf
graphviz